Utility refactor and a test
